Moved Wallet into the shared crate and parameterized it by its interactive components.
