+++ apps/src/bin/anoma-wallet/cli.rs
@@ -12,8 +12,10 @@ use namada::types::masp::{MaspValue, PaymentAddress};
 use namada_apps::cli;
 use namada_apps::cli::{args, cmds, Context};
 use namada::ledger::masp::find_valid_diversifier;
-use namada_apps::wallet::{DecryptionError, FindKeyError};
+use namada_apps::wallet::DecryptionError;
 use rand_core::OsRng;
+use namada_apps::wallet::CliWalletUtils;
+use namada::ledger::wallet::FindKeyError;
 
 pub fn main() -> Result<()> {
     let (cmd, mut ctx) = cli::anoma_wallet_cli()?;
@@ -80,7 +82,7 @@ fn address_key_find(
         println!("Viewing key: {}", viewing_key);
         if unsafe_show_secret {
             // Check if alias is also a spending key
-            match wallet.find_spending_key(&alias) {
+            match wallet.find_spending_key::<CliWalletUtils>(&alias) {
                 Ok(spending_key) => println!("Spending key: {}", spending_key),
                 Err(FindKeyError::KeyNotFound) => {}
                 Err(err) => eprintln!("{}", err),
@@ -142,7 +144,7 @@ fn spending_keys_list(
             // Print those too if they are available and requested.
             if unsafe_show_secret {
                 if let Some(spending_key) = spending_key_opt {
-                    match spending_key.get(decrypt, None) {
+                    match spending_key.get::<CliWalletUtils>(decrypt, None) {
                         // Here the spending key is unencrypted or successfully
                         // decrypted
                         Ok(spending_key) => {
@@ -200,7 +202,7 @@ fn spending_key_gen(
 ) {
     let mut wallet = ctx.wallet;
     let alias = alias.to_lowercase();
-    let (alias, _key) = wallet.gen_spending_key(alias, unsafe_dont_encrypt);
+    let (alias, _key) = wallet.gen_spending_key::<CliWalletUtils>(alias, unsafe_dont_encrypt);
     namada_apps::wallet::save(&wallet).unwrap_or_else(|err| eprintln!("{}", err));
     println!(
         "Successfully added a spending key with alias: \"{}\"",
@@ -228,7 +230,7 @@ fn payment_address_gen(
         .expect("a PaymentAddress");
     let mut wallet = ctx.wallet;
     let alias = wallet
-        .insert_payment_addr(
+        .insert_payment_addr::<CliWalletUtils>(
             alias,
             PaymentAddress::from(payment_addr).pinned(pin),
         )
@@ -257,7 +259,7 @@ fn address_key_add(
         MaspValue::FullViewingKey(viewing_key) => {
             let alias = ctx
                 .wallet
-                .insert_viewing_key(alias, viewing_key)
+                .insert_viewing_key::<CliWalletUtils>(alias, viewing_key)
                 .unwrap_or_else(|| {
                     eprintln!("Viewing key not added");
                     cli::safe_exit(1);
@@ -267,7 +269,7 @@ fn address_key_add(
         MaspValue::ExtendedSpendingKey(spending_key) => {
             let alias = ctx
                 .wallet
-                .encrypt_insert_spending_key(
+                .encrypt_insert_spending_key::<CliWalletUtils>(
                     alias,
                     spending_key,
                     unsafe_dont_encrypt,
@@ -281,7 +283,7 @@ fn address_key_add(
         MaspValue::PaymentAddress(payment_addr) => {
             let alias = ctx
                 .wallet
-                .insert_payment_addr(alias, payment_addr)
+                .insert_payment_addr::<CliWalletUtils>(alias, payment_addr)
                 .unwrap_or_else(|| {
                     eprintln!("Payment address not added");
                     cli::safe_exit(1);
@@ -307,7 +309,7 @@ fn key_and_address_gen(
     }: args::KeyAndAddressGen,
 ) {
     let mut wallet = ctx.wallet;
-    let (alias, _key) = wallet.gen_key(scheme, alias, unsafe_dont_encrypt);
+    let (alias, _key) = wallet.gen_key::<CliWalletUtils>(scheme, alias, unsafe_dont_encrypt);
     namada_apps::wallet::save(&wallet).unwrap_or_else(|err| eprintln!("{}", err));
     println!(
         "Successfully added a key and an address with alias: \"{}\"",
@@ -327,7 +329,7 @@ fn key_find(
 ) {
     let mut wallet = ctx.wallet;
     let found_keypair = match public_key {
-        Some(pk) => wallet.find_key_by_pk(&pk),
+        Some(pk) => wallet.find_key_by_pk::<CliWalletUtils>(&pk),
         None => {
             let alias = alias.or(value);
             match alias {
@@ -338,7 +340,7 @@ fn key_find(
                     );
                     cli::safe_exit(1)
                 }
-                Some(alias) => wallet.find_key(alias.to_lowercase()),
+                Some(alias) => wallet.find_key::<CliWalletUtils>(alias.to_lowercase()),
             }
         }
     };
@@ -386,7 +388,7 @@ fn key_list(
             if let Some(pkh) = pkh {
                 writeln!(w, "    Public key hash: {}", pkh).unwrap();
             }
-            match stored_keypair.get(decrypt, None) {
+            match stored_keypair.get::<CliWalletUtils>(decrypt, None) {
                 Ok(keypair) => {
                     writeln!(w, "    Public key: {}", keypair.ref_to())
                         .unwrap();
@@ -410,7 +412,7 @@ fn key_list(
 fn key_export(ctx: Context, args::KeyExport { alias }: args::KeyExport) {
     let mut wallet = ctx.wallet;
     wallet
-        .find_key(alias.to_lowercase())
+        .find_key::<CliWalletUtils>(alias.to_lowercase())
         .map(|keypair| {
             let file_data = keypair
                 .try_to_vec()
@@ -483,7 +485,7 @@ fn address_or_alias_find(ctx: Context, args: args::AddressOrAliasFind) {
 fn address_add(ctx: Context, args: args::AddressAdd) {
     let mut wallet = ctx.wallet;
     if wallet
-        .add_address(args.alias.clone().to_lowercase(), args.address)
+        .add_address::<CliWalletUtils>(args.alias.clone().to_lowercase(), args.address)
         .is_none()
     {
         eprintln!("Address not added");
