+++ apps/src/lib/client/signing.rs
@@ -13,11 +13,12 @@ use std::path::PathBuf;
 use super::rpc;
 use crate::cli::{self, args};
 use crate::client::tendermint_rpc_types::TxBroadcastData;
-use crate::wallet::Wallet;
+use namada::ledger::wallet::Wallet;
+use namada::ledger::wallet::WalletUtils;
 
 /// Find the public key for the given address and try to load the keypair
 /// for it from the wallet. Panics if the key cannot be found or loaded.
-pub async fn find_keypair(
+pub async fn find_keypair<U: WalletUtils>(
     client: &HttpClient,
     wallet: &mut Wallet<PathBuf>,
     addr: &Address,
@@ -37,7 +38,7 @@ pub async fn find_keypair(
                     );
                     cli::safe_exit(1);
                 });
-            wallet.find_key_by_pk(&public_key).unwrap_or_else(|err| {
+            wallet.find_key_by_pk::<U>(&public_key).unwrap_or_else(|err| {
                 eprintln!(
                     "Unable to load the keypair from the wallet for public \
                      key {}. Failed with: {}",
@@ -47,7 +48,7 @@ pub async fn find_keypair(
             })
         }
         Address::Implicit(ImplicitAddress(pkh)) => {
-            wallet.find_key_by_pkh(pkh).unwrap_or_else(|err| {
+            wallet.find_key_by_pkh::<U>(pkh).unwrap_or_else(|err| {
                 eprintln!(
                     "Unable to load the keypair from the wallet for the \
                      implicit address {}. Failed with: {}",
@@ -85,7 +86,7 @@ pub enum TxSigningKey {
 /// signer. Return the given signing key or public key of the given signer if
 /// possible. If no explicit signer given, use the `default`. If no `default`
 /// is given, panics.
-pub async fn tx_signer(
+pub async fn tx_signer<U: WalletUtils>(
     client: &HttpClient,
     wallet: &mut Wallet<PathBuf>,
     args: &args::Tx,
@@ -104,7 +105,7 @@ pub async fn tx_signer(
         }
         TxSigningKey::WalletAddress(signer) => {
             let signer = signer;
-            let signing_key = find_keypair(
+            let signing_key = find_keypair::<U>(
                 client,
                 wallet,
                 &signer,
@@ -114,14 +115,14 @@ pub async fn tx_signer(
             // PK first
             if matches!(signer, Address::Implicit(_)) {
                 let pk: common::PublicKey = signing_key.ref_to();
-                super::tx::reveal_pk_if_needed(client, wallet, &pk, args).await;
+                super::tx::reveal_pk_if_needed::<U>(client, wallet, &pk, args).await;
             }
             signing_key
         }
         TxSigningKey::SecretKey(signing_key) => {
             // Check if the signing key needs to reveal its PK first
             let pk: common::PublicKey = signing_key.ref_to();
-            super::tx::reveal_pk_if_needed(client, wallet, &pk, args).await;
+            super::tx::reveal_pk_if_needed::<U>(client, wallet, &pk, args).await;
             signing_key
         }
         TxSigningKey::None => {
@@ -141,14 +142,14 @@ pub async fn tx_signer(
 /// hashes needed for monitoring the tx on chain.
 ///
 /// If it is a dry run, it is not put in a wrapper, but returned as is.
-pub async fn sign_tx(
+pub async fn sign_tx<U: WalletUtils>(
     client: &HttpClient,
     wallet: &mut Wallet<PathBuf>,
     tx: Tx,
     args: &args::Tx,
     default: TxSigningKey,
 ) -> TxBroadcastData {
-    let keypair = tx_signer(client, wallet, args, default).await;
+    let keypair = tx_signer::<U>(client, wallet, args, default).await;
     let tx = tx.sign(&keypair);
 
     let epoch = rpc::query_epoch(client)
