+++ apps/src/lib/wallet/mod.rs
@@ -19,426 +19,173 @@ use namada::types::masp::{
 };
 pub use store::wallet_file;
 use thiserror::Error;
+use namada::ledger::wallet::ConfirmationResponse;
 
-use self::alias::Alias;
-pub use self::keys::{DecryptionError, StoredKeypair};
-use self::store::Store;
-pub use self::store::{ValidatorData, ValidatorKeys};
+pub use namada::ledger::wallet::{DecryptionError, StoredKeypair};
+use namada::ledger::wallet::{Store, Wallet};
+pub use namada::ledger::wallet::{ValidatorData, ValidatorKeys};
 use crate::cli;
 use crate::config::genesis::genesis_config::GenesisConfig;
+use namada::ledger::wallet::{WalletUtils, Alias};
+use std::io::prelude::*;
+use std::io::{self, Write};
+use namada::ledger::wallet::FindKeyError;
 
-#[derive(Debug)]
-pub struct Wallet<C> {
-    store_dir: C,
-    store: Store,
-    decrypted_key_cache: HashMap<Alias, common::SecretKey>,
-    decrypted_spendkey_cache: HashMap<Alias, ExtendedSpendingKey>,
-}
+pub struct CliWalletUtils;
 
-#[derive(Error, Debug)]
-pub enum FindKeyError {
-    #[error("No matching key found")]
-    KeyNotFound,
-    #[error("{0}")]
-    KeyDecryptionError(keys::DecryptionError),
-}
-
-impl<C> Wallet<C> {
-    /// Add addresses from a genesis configuration.
-    pub fn add_genesis_addresses(&mut self, genesis: GenesisConfig) {
-        self.store.add_genesis_addresses(genesis)
-    }
-
-    /// Generate a new keypair and derive an implicit address from its public
-    /// and insert them into the store with the provided alias, converted to
-    /// lower case. If none provided, the alias will be the public key hash (in
-    /// lowercase too). If the key is to be encrypted, will prompt for
-    /// password from stdin. Stores the key in decrypted key cache and
-    /// returns the alias of the key and a reference-counting pointer to the
-    /// key.
-    pub fn gen_key(
-        &mut self,
-        scheme: SchemeType,
-        alias: Option<String>,
-        unsafe_dont_encrypt: bool,
-    ) -> (String, common::SecretKey) {
-        let password = read_and_confirm_pwd(unsafe_dont_encrypt);
-        let (alias, key) = self.store.gen_key(scheme, alias, password);
-        // Cache the newly added key
-        self.decrypted_key_cache.insert(alias.clone(), key.clone());
-        (alias.into(), key)
-    }
-
-    pub fn gen_spending_key(
-        &mut self,
-        alias: String,
-        unsafe_dont_encrypt: bool,
-    ) -> (String, ExtendedSpendingKey) {
-        let password = new_password_prompt(unsafe_dont_encrypt);
-        let (alias, key) = self.store.gen_spending_key(alias, password);
-        // Cache the newly added key
-        self.decrypted_spendkey_cache.insert(alias.clone(), key);
-        (alias.into(), key)
-    }
-
-    /// Generate keypair
-    /// for signing protocol txs and for the DKG (which will also be stored)
-    /// A protocol keypair may be optionally provided, indicating that
-    /// we should re-use a keypair already in the wallet
-    pub fn gen_validator_keys(
-        &mut self,
-        protocol_pk: Option<common::PublicKey>,
-        scheme: SchemeType,
-    ) -> Result<ValidatorKeys, FindKeyError> {
-        let protocol_keypair = protocol_pk.map(|pk| {
-            self.find_key_by_pkh(&PublicKeyHash::from(&pk))
-                .ok()
-                .or_else(|| {
-                    self.store
-                        .validator_data
-                        .take()
-                        .map(|data| data.keys.protocol_keypair)
-                })
-                .ok_or(FindKeyError::KeyNotFound)
-        });
-        match protocol_keypair {
-            Some(Err(err)) => Err(err),
-            other => Ok(Store::gen_validator_keys(
-                other.map(|res| res.unwrap()),
-                scheme,
-            )),
+impl WalletUtils for CliWalletUtils {
+    /// Prompt for pssword and confirm it if parameter is false
+    fn new_password_prompt(unsafe_dont_encrypt: bool) -> Option<String> {
+        let password = if unsafe_dont_encrypt {
+            println!("Warning: The keypair will NOT be encrypted.");
+            None
+        } else {
+            Some(Self::read_password("Enter your encryption password: "))
+        };
+        // Bis repetita for confirmation.
+        let pwd = if unsafe_dont_encrypt {
+            None
+        } else {
+            Some(Self::read_password(
+                "To confirm, please enter the same encryption password once \
+                 more: ",
+            ))
+        };
+        if pwd != password {
+            eprintln!("Your two inputs do not match!");
+            cli::safe_exit(1)
         }
+        password
     }
 
-    /// Add validator data to the store
-    pub fn add_validator_data(
-        &mut self,
-        address: Address,
-        keys: ValidatorKeys,
-    ) {
-        self.store.add_validator_data(address, keys);
-    }
-
-    /// Returns the validator data, if it exists.
-    pub fn get_validator_data(&self) -> Option<&ValidatorData> {
-        self.store.get_validator_data()
-    }
-
-    /// Returns the validator data, if it exists.
-    /// [`Wallet::save`] cannot be called after using this
-    /// method as it involves a partial move
-    pub fn take_validator_data(self) -> Option<ValidatorData> {
-        self.store.validator_data()
-    }
-
-    /// Find the stored key by an alias, a public key hash or a public key.
-    /// If the key is encrypted, will prompt for password from stdin.
-    /// Any keys that are decrypted are stored in and read from a cache to avoid
-    /// prompting for password multiple times.
-    pub fn find_key(
-        &mut self,
-        alias_pkh_or_pk: impl AsRef<str>,
-    ) -> Result<common::SecretKey, FindKeyError> {
-        // Try cache first
-        if let Some(cached_key) = self
-            .decrypted_key_cache
-            .get(&alias_pkh_or_pk.as_ref().into())
-        {
-            return Ok(cached_key.clone());
+    /// Read the password for encryption from the file/env/stdin with confirmation.
+    fn read_and_confirm_pwd(unsafe_dont_encrypt: bool) -> Option<String> {
+        let password = if unsafe_dont_encrypt {
+            println!("Warning: The keypair will NOT be encrypted.");
+            None
+        } else {
+            Some(Self::read_password("Enter your encryption password: "))
+        };
+        // Bis repetita for confirmation.
+        let to_confirm = if unsafe_dont_encrypt {
+            None
+        } else {
+            Some(Self::read_password(
+                "To confirm, please enter the same encryption password once more: ",
+            ))
+        };
+        if to_confirm != password {
+            eprintln!("Your two inputs do not match!");
+            cli::safe_exit(1)
         }
-        // If not cached, look-up in store
-        let stored_key = self
-            .store
-            .find_key(alias_pkh_or_pk.as_ref())
-            .ok_or(FindKeyError::KeyNotFound)?;
-        Self::decrypt_stored_key(
-            &mut self.decrypted_key_cache,
-            stored_key,
-            alias_pkh_or_pk.into(),
-        )
+        password
     }
 
-    pub fn find_spending_key(
-        &mut self,
-        alias: impl AsRef<str>,
-    ) -> Result<ExtendedSpendingKey, FindKeyError> {
-        // Try cache first
-        if let Some(cached_key) =
-            self.decrypted_spendkey_cache.get(&alias.as_ref().into())
-        {
-            return Ok(*cached_key);
+    /// Read the password for encryption/decryption from the file/env/stdin. Panics
+    /// if all options are empty/invalid.
+    fn read_password(prompt_msg: &str) -> String {
+        let pwd = match env::var("ANOMA_WALLET_PASSWORD_FILE") {
+            Ok(path) => fs::read_to_string(path)
+                .expect("Something went wrong reading the file"),
+            Err(_) => match env::var("ANOMA_WALLET_PASSWORD") {
+                Ok(password) => password,
+                Err(_) => rpassword::read_password_from_tty(Some(prompt_msg))
+                    .unwrap_or_default(),
+            },
+        };
+        if pwd.is_empty() {
+            eprintln!("Password cannot be empty");
+            cli::safe_exit(1)
         }
-        // If not cached, look-up in store
-        let stored_spendkey = self
-            .store
-            .find_spending_key(alias.as_ref())
-            .ok_or(FindKeyError::KeyNotFound)?;
-        Self::decrypt_stored_key(
-            &mut self.decrypted_spendkey_cache,
-            stored_spendkey,
-            alias.into(),
-        )
+        pwd
     }
 
-    pub fn find_viewing_key(
-        &mut self,
-        alias: impl AsRef<str>,
-    ) -> Result<&ExtendedViewingKey, FindKeyError> {
-        self.store
-            .find_viewing_key(alias.as_ref())
-            .ok_or(FindKeyError::KeyNotFound)
-    }
+    /// The given alias has been selected but conflicts with another alias in
+    /// the store. Offer the user to either replace existing mapping, alter the
+    /// chosen alias to a name of their chosing, or cancel the aliasing.
+    fn show_overwrite_confirmation(
+        alias: &Alias,
+        alias_for: &str,
+    ) -> ConfirmationResponse {
+        print!(
+            "You're trying to create an alias \"{}\" that already exists for {} \
+             in your store.\nWould you like to replace it? \
+             s(k)ip/re(p)lace/re(s)elect: ",
+            alias, alias_for
+        );
+        io::stdout().flush().unwrap();
 
-    pub fn find_payment_addr(
-        &self,
-        alias: impl AsRef<str>,
-    ) -> Option<&PaymentAddress> {
-        self.store.find_payment_addr(alias.as_ref())
-    }
-
-    /// Find the stored key by a public key.
-    /// If the key is encrypted, will prompt for password from stdin.
-    /// Any keys that are decrypted are stored in and read from a cache to avoid
-    /// prompting for password multiple times.
-    pub fn find_key_by_pk(
-        &mut self,
-        pk: &common::PublicKey,
-    ) -> Result<common::SecretKey, FindKeyError> {
-        // Try to look-up alias for the given pk. Otherwise, use the PKH string.
-        let pkh: PublicKeyHash = pk.into();
-        let alias = self
-            .store
-            .find_alias_by_pkh(&pkh)
-            .unwrap_or_else(|| pkh.to_string().into());
-        // Try read cache
-        if let Some(cached_key) = self.decrypted_key_cache.get(&alias) {
-            return Ok(cached_key.clone());
-        }
-        // Look-up from store
-        let stored_key = self
-            .store
-            .find_key_by_pk(pk)
-            .ok_or(FindKeyError::KeyNotFound)?;
-        Self::decrypt_stored_key(
-            &mut self.decrypted_key_cache,
-            stored_key,
-            alias,
-        )
-    }
-
-    /// Find the stored key by a public key hash.
-    /// If the key is encrypted, will prompt for password from stdin.
-    /// Any keys that are decrypted are stored in and read from a cache to avoid
-    /// prompting for password multiple times.
-    pub fn find_key_by_pkh(
-        &mut self,
-        pkh: &PublicKeyHash,
-    ) -> Result<common::SecretKey, FindKeyError> {
-        // Try to look-up alias for the given pk. Otherwise, use the PKH string.
-        let alias = self
-            .store
-            .find_alias_by_pkh(pkh)
-            .unwrap_or_else(|| pkh.to_string().into());
-        // Try read cache
-        if let Some(cached_key) = self.decrypted_key_cache.get(&alias) {
-            return Ok(cached_key.clone());
-        }
-        // Look-up from store
-        let stored_key = self
-            .store
-            .find_key_by_pkh(pkh)
-            .ok_or(FindKeyError::KeyNotFound)?;
-        Self::decrypt_stored_key(
-            &mut self.decrypted_key_cache,
-            stored_key,
-            alias,
-        )
-    }
-
-    /// Decrypt stored key, if it's not stored un-encrypted.
-    /// If a given storage key needs to be decrypted, prompt for password from
-    /// stdin and if successfully decrypted, store it in a cache.
-    fn decrypt_stored_key<
-        T: FromStr + Display + BorshSerialize + BorshDeserialize + Clone,
-    >(
-        decrypted_key_cache: &mut HashMap<Alias, T>,
-        stored_key: &StoredKeypair<T>,
-        alias: Alias,
-    ) -> Result<T, FindKeyError>
-    where
-        <T as std::str::FromStr>::Err: Display,
-    {
-        match stored_key {
-            StoredKeypair::Encrypted(encrypted) => {
-                let password = read_password("Enter decryption password: ");
-                let key = encrypted
-                    .decrypt(password)
-                    .map_err(FindKeyError::KeyDecryptionError)?;
-                decrypted_key_cache.insert(alias.clone(), key);
-                decrypted_key_cache
-                    .get(&alias)
-                    .cloned()
-                    .ok_or(FindKeyError::KeyNotFound)
+        let mut buffer = String::new();
+        // Get the user to select between 3 choices
+        match io::stdin().read_line(&mut buffer) {
+            Ok(size) if size > 0 => {
+                // Isolate the single character representing the choice
+                let byte = buffer.chars().next().unwrap();
+                buffer.clear();
+                match byte {
+                    'p' | 'P' => return ConfirmationResponse::Replace,
+                    's' | 'S' => {
+                        // In the case of reselection, elicit new alias
+                        print!("Please enter a different alias: ");
+                        io::stdout().flush().unwrap();
+                        if io::stdin().read_line(&mut buffer).is_ok() {
+                            return ConfirmationResponse::Reselect(
+                                buffer.trim().into(),
+                            );
+                        }
+                    }
+                    'k' | 'K' => return ConfirmationResponse::Skip,
+                    // Input is senseless fall through to repeat prompt
+                    _ => {}
+                };
             }
-            StoredKeypair::Raw(raw) => Ok(raw.clone()),
+            _ => {}
         }
+        // Input is senseless fall through to repeat prompt
+        println!("Invalid option, try again.");
+        Self::show_overwrite_confirmation(alias, alias_for)
     }
+}
 
-    /// Get all known keys by their alias, paired with PKH, if known.
-    pub fn get_keys(
-        &self,
-    ) -> HashMap<
-        String,
-        (&StoredKeypair<common::SecretKey>, Option<&PublicKeyHash>),
-    > {
-        self.store
-            .get_keys()
-            .into_iter()
-            .map(|(alias, value)| (alias.into(), value))
-            .collect()
+/// Generate keypair
+/// for signing protocol txs and for the DKG (which will also be stored)
+/// A protocol keypair may be optionally provided, indicating that
+/// we should re-use a keypair already in the wallet
+pub fn gen_validator_keys(
+    wallet: &mut Wallet<PathBuf>,
+    protocol_pk: Option<common::PublicKey>,
+    scheme: SchemeType,
+) -> Result<ValidatorKeys, FindKeyError> {
+    let protocol_keypair = protocol_pk.map(|pk| {
+        wallet.find_key_by_pkh::<CliWalletUtils>(&PublicKeyHash::from(&pk))
+            .ok()
+            .or_else(|| {
+                wallet.store_mut()
+                    .validator_data()
+                    .take()
+                    .map(|data| data.keys.protocol_keypair.clone())
+            })
+            .ok_or(FindKeyError::KeyNotFound)
+    });
+    match protocol_keypair {
+        Some(Err(err)) => Err(err),
+        other => Ok(store::gen_validator_keys(
+            other.map(|res| res.unwrap()),
+            scheme,
+        )),
     }
+}
 
-    /// Find the stored address by an alias.
-    pub fn find_address(&self, alias: impl AsRef<str>) -> Option<&Address> {
-        self.store.find_address(alias)
-    }
-
-    /// Find an alias by the address if it's in the wallet.
-    pub fn find_alias(&self, address: &Address) -> Option<&Alias> {
-        self.store.find_alias(address)
-    }
-
-    /// Get all known addresses by their alias, paired with PKH, if known.
-    pub fn get_addresses(&self) -> HashMap<String, Address> {
-        self.store
-            .get_addresses()
-            .iter()
-            .map(|(alias, value)| (alias.into(), value.clone()))
-            .collect()
-    }
-
-    /// Get all known payment addresses by their alias
-    pub fn get_payment_addrs(&self) -> HashMap<String, PaymentAddress> {
-        self.store
-            .get_payment_addrs()
-            .iter()
-            .map(|(alias, value)| (alias.into(), *value))
-            .collect()
-    }
-
-    /// Get all known viewing keys by their alias
-    pub fn get_viewing_keys(&self) -> HashMap<String, ExtendedViewingKey> {
-        self.store
-            .get_viewing_keys()
-            .iter()
-            .map(|(alias, value)| (alias.into(), *value))
-            .collect()
-    }
-
-    /// Get all known viewing keys by their alias
-    pub fn get_spending_keys(
-        &self,
-    ) -> HashMap<String, &StoredKeypair<ExtendedSpendingKey>> {
-        self.store
-            .get_spending_keys()
-            .iter()
-            .map(|(alias, value)| (alias.into(), value))
-            .collect()
-    }
-
-    /// Add a new address with the given alias. If the alias is already used,
-    /// will ask whether the existing alias should be replaced, a different
-    /// alias is desired, or the alias creation should be cancelled. Return
-    /// the chosen alias if the address has been added, otherwise return
-    /// nothing.
-    pub fn add_address(
-        &mut self,
-        alias: impl AsRef<str>,
-        address: Address,
-    ) -> Option<String> {
-        self.store
-            .insert_address(alias.into(), address)
-            .map(Into::into)
-    }
-
-    /// Insert a new key with the given alias. If the alias is already used,
-    /// will prompt for overwrite confirmation.
-    pub fn insert_keypair(
-        &mut self,
-        alias: String,
-        keypair: StoredKeypair<common::SecretKey>,
-        pkh: PublicKeyHash,
-    ) -> Option<String> {
-        self.store
-            .insert_keypair(alias.into(), keypair, pkh)
-            .map(Into::into)
-    }
-
-    pub fn insert_viewing_key(
-        &mut self,
-        alias: String,
-        view_key: ExtendedViewingKey,
-    ) -> Option<String> {
-        self.store
-            .insert_viewing_key(alias.into(), view_key)
-            .map(Into::into)
-    }
-
-    pub fn insert_spending_key(
-        &mut self,
-        alias: String,
-        spend_key: StoredKeypair<ExtendedSpendingKey>,
-        viewkey: ExtendedViewingKey,
-    ) -> Option<String> {
-        self.store
-            .insert_spending_key(alias.into(), spend_key, viewkey)
-            .map(Into::into)
-    }
-
-    pub fn encrypt_insert_spending_key(
-        &mut self,
-        alias: String,
-        spend_key: ExtendedSpendingKey,
-        unsafe_dont_encrypt: bool,
-    ) -> Option<String> {
-        let password = new_password_prompt(unsafe_dont_encrypt);
-        self.store
-            .insert_spending_key(
-                alias.into(),
-                StoredKeypair::new(spend_key, password).0,
-                ExtendedFullViewingKey::from(&spend_key.into()).into(),
-            )
-            .map(Into::into)
-    }
-
-    pub fn insert_payment_addr(
-        &mut self,
-        alias: String,
-        payment_addr: PaymentAddress,
-    ) -> Option<String> {
-        self.store
-            .insert_payment_addr(alias.into(), payment_addr)
-            .map(Into::into)
-    }
-
-    /// Extend this wallet from pre-genesis validator wallet.
-    pub fn extend_from_pre_genesis_validator(
-        &mut self,
-        validator_address: Address,
-        validator_alias: Alias,
-        other: pre_genesis::ValidatorWallet,
-    ) {
-        self.store.extend_from_pre_genesis_validator(
-            validator_address,
-            validator_alias,
-            other,
-        )
+/// Add addresses from a genesis configuration.
+pub fn add_genesis_addresses(wallet: &mut Wallet<PathBuf>, genesis: GenesisConfig) {
+    for (alias, addr) in defaults::addresses_from_genesis(genesis) {
+        wallet.add_address::<CliWalletUtils>(alias.normalize(), addr);
     }
 }
 
 /// Save the wallet store to a file.
 pub fn save(wallet: &Wallet<PathBuf>) -> std::io::Result<()> {
-    self::store::save(&wallet.store, &wallet.store_dir)
+    self::store::save(&wallet.store(), &wallet.store_dir())
 }
 
 /// Load a wallet from the store file.
@@ -447,12 +194,7 @@ pub fn load(store_dir: &Path) -> Option<Wallet<PathBuf>> {
         eprintln!("Unable to load the wallet: {}", err);
         cli::safe_exit(1)
     });
-    Some(Wallet::<PathBuf> {
-        store_dir: store_dir.to_path_buf(),
-        store,
-        decrypted_key_cache: HashMap::default(),
-        decrypted_spendkey_cache: HashMap::default(),
-    })
+    Some(Wallet::<PathBuf>::new(store_dir.to_path_buf(), store))
 }
 
 /// Load a wallet from the store file or create a new wallet without any
@@ -462,12 +204,7 @@ pub fn load_or_new(store_dir: &Path) -> Wallet<PathBuf> {
         eprintln!("Unable to load the wallet: {}", err);
         cli::safe_exit(1)
     });
-    Wallet::<PathBuf> {
-        store_dir: store_dir.to_path_buf(),
-        store,
-        decrypted_key_cache: HashMap::default(),
-        decrypted_spendkey_cache: HashMap::default(),
-    }
+    Wallet::<PathBuf>::new(store_dir.to_path_buf(), store)
 }
 
 /// Load a wallet from the store file or create a new one with the default
@@ -481,76 +218,5 @@ pub fn load_or_new_from_genesis(
             eprintln!("Unable to load the wallet: {}", err);
             cli::safe_exit(1)
         });
-    Wallet::<PathBuf> {
-        store_dir: store_dir.to_path_buf(),
-        store,
-        decrypted_key_cache: HashMap::default(),
-        decrypted_spendkey_cache: HashMap::default(),
-    }
-}
-
-/// Prompt for pssword and confirm it if parameter is false
-fn new_password_prompt(unsafe_dont_encrypt: bool) -> Option<String> {
-    let password = if unsafe_dont_encrypt {
-        println!("Warning: The keypair will NOT be encrypted.");
-        None
-    } else {
-        Some(read_password("Enter your encryption password: "))
-    };
-    // Bis repetita for confirmation.
-    let pwd = if unsafe_dont_encrypt {
-        None
-    } else {
-        Some(read_password(
-            "To confirm, please enter the same encryption password once \
-             more: ",
-        ))
-    };
-    if pwd != password {
-        eprintln!("Your two inputs do not match!");
-        cli::safe_exit(1)
-    }
-    password
-}
-
-/// Read the password for encryption from the file/env/stdin with confirmation.
-pub fn read_and_confirm_pwd(unsafe_dont_encrypt: bool) -> Option<String> {
-    let password = if unsafe_dont_encrypt {
-        println!("Warning: The keypair will NOT be encrypted.");
-        None
-    } else {
-        Some(read_password("Enter your encryption password: "))
-    };
-    // Bis repetita for confirmation.
-    let to_confirm = if unsafe_dont_encrypt {
-        None
-    } else {
-        Some(read_password(
-            "To confirm, please enter the same encryption password once more: ",
-        ))
-    };
-    if to_confirm != password {
-        eprintln!("Your two inputs do not match!");
-        cli::safe_exit(1)
-    }
-    password
-}
-
-/// Read the password for encryption/decryption from the file/env/stdin. Panics
-/// if all options are empty/invalid.
-pub fn read_password(prompt_msg: &str) -> String {
-    let pwd = match env::var("ANOMA_WALLET_PASSWORD_FILE") {
-        Ok(path) => fs::read_to_string(path)
-            .expect("Something went wrong reading the file"),
-        Err(_) => match env::var("ANOMA_WALLET_PASSWORD") {
-            Ok(password) => password,
-            Err(_) => rpassword::read_password_from_tty(Some(prompt_msg))
-                .unwrap_or_default(),
-        },
-    };
-    if pwd.is_empty() {
-        eprintln!("Password cannot be empty");
-        cli::safe_exit(1)
-    }
-    pwd
+    Wallet::<PathBuf>::new(store_dir.to_path_buf(), store)
 }
