+++ shared/Cargo.toml
@@ -107,6 +107,7 @@ prost = "0.9.0"
 pwasm-utils = {version = "0.18.0", optional = true}
 rayon = {version = "=1.5.3", optional = true}
 rust_decimal = "1.26.1"
+serde = {version = "1.0.125", features = ["derive"]}
 serde_json = "1.0.62"
 sha2 = "0.9.3"
 # We switch off "blake2b" because it cannot be compiled to wasm
@@ -132,6 +133,9 @@ masp_proofs = { git = "https://github.com/anoma/masp", rev = "bee40fc465f6afbd10
 rand = {version = "0.8", default-features = false, optional = true}
 rand_core = {version = "0.6", default-features = false, optional = true}
 zeroize = "1.5.5"
+toml = "0.5.8"
+bimap = {version = "0.6.2", features = ["serde"]}
+orion = "0.16.0"
 
 [dev-dependencies]
 assert_matches = "1.5.0"
